+++ movingspeed.h
@@ -5,15 +5,15 @@
 class MovingSpeed
 {
 public:
-    static float x_speed;
-    static float y_speed;
+    static long double x_speed;
+    static long double y_speed;
 
     static int current_power;
-    static float current_speed;
+    static double current_speed;
 
     const static int division_power_factor_index;
     const static int max_power;
-    const static float division_factor_speed;
+    const static double division_factor_speed;
     const static int max_speed[];
 };
